+++ src/polyvec.rs
@@ -3,9 +3,10 @@
 use crate::params::{KyberMode, N, POLYBYTES, Q32};
 use crate::poly::Poly;
 use alloc::vec::Vec;
+use zeroize::Zeroize;
 
 /// A vector of K polynomials.
-#[derive(Clone)]
+#[derive(Clone, Zeroize)]
 pub struct PolyVec {
     pub vec: Vec<Poly>,
 }
@@ -80,11 +81,11 @@ impl PolyVec {
                 for i in 0..k {
                     for j in 0..N / 8 {
                         let mut t = [0u16; 8];
-                        for m in 0..8 {
+                        for (m, tm) in t.iter_mut().enumerate() {
                             let mut u = self.vec[i].coeffs[8 * j + m];
                             u += (u >> 15) & (Q32 as i16);
-                            t[m] =
-                                ((((u as u32) << 11) + Q32 as u32 / 2) / Q32 as u32 & 0x7FF) as u16;
+                            *tm =
+                                (((((u as u32) << 11) + Q32 as u32 / 2) / Q32 as u32) & 0x7FF) as u16;
                         }
                         r[idx] = t[0] as u8;
                         r[idx + 1] = ((t[0] >> 8) | (t[1] << 3)) as u8;
@@ -107,11 +108,11 @@ impl PolyVec {
                 for i in 0..k {
                     for j in 0..N / 4 {
                         let mut t = [0u16; 4];
-                        for m in 0..4 {
+                        for (m, tm) in t.iter_mut().enumerate() {
                             let mut u = self.vec[i].coeffs[4 * j + m];
                             u += (u >> 15) & (Q32 as i16);
-                            t[m] =
-                                ((((u as u32) << 10) + Q32 as u32 / 2) / Q32 as u32 & 0x3FF) as u16;
+                            *tm =
+                                (((((u as u32) << 10) + Q32 as u32 / 2) / Q32 as u32) & 0x3FF) as u16;
                         }
                         r[idx] = t[0] as u8;
                         r[idx + 1] = ((t[0] >> 8) | (t[1] << 2)) as u8;
@@ -156,9 +157,9 @@ impl PolyVec {
                         t[7] =
                             ((u16::from(a[idx + 9]) >> 5) | (u16::from(a[idx + 10]) << 3)) & 0x7FF;
                         idx += 11;
-                        for m in 0..8 {
+                        for (m, &tm) in t.iter().enumerate() {
                             pv.vec[i].coeffs[8 * j + m] =
-                                (((t[m] as u32) * Q32 as u32 + 1024) >> 11) as i16;
+                                (((tm as u32) * Q32 as u32 + 1024) >> 11) as i16;
                         }
                     }
                 }
@@ -176,9 +177,9 @@ impl PolyVec {
                         t[3] =
                             ((u16::from(a[idx + 3]) >> 6) | (u16::from(a[idx + 4]) << 2)) & 0x3FF;
                         idx += 5;
-                        for m in 0..4 {
+                        for (m, &tm) in t.iter().enumerate() {
                             pv.vec[i].coeffs[4 * j + m] =
-                                (((t[m] as u32) * Q32 as u32 + 512) >> 10) as i16;
+                                (((tm as u32) * Q32 as u32 + 512) >> 10) as i16;
                         }
                     }
                 }